Add new HostFacts data structure and refactor existing structs
Using this new setup, data structures for metric collection are
split in two: one data structure for metrics which are expected to
change during the runtime of this program and one for which metrics
are exprected to stay the same during the runtime of this program.
These are named LiveMetrics and HostFacts respectively.

Once a remote endpoint is implemented, the HostFacts struct will be
send once on startup, while the LiveMetrics struct will be send
continuously.